% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{Dataset0}
\alias{Dataset0}
\title{Dataset0}
\usage{
Dataset0(size = 2000)
}
\arguments{
\item{size}{[integer] length of dataset}
}
\value{
[list(t,X,Y)] Dataset Y, with covariate X and time axis t
}
\description{
Generate a test dataset, Y is a Gaussian law where location move with time, but scale is fixed at 0.2, X is the location parameter
}
\examples{
## Data
size = 2000
data = Dataset0(size)
t = data$t  ## Time axis
X = data$X  ## Covariate
Y = data$Y  ## Dataset to fit
}
