% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{Dataset2}
\alias{Dataset2}
\title{Dataset2}
\usage{
Dataset2(size = 2000, alpha = 0.05, beta = 0.6)
}
\arguments{
\item{size}{[integer] length of dataset}

\item{alpha}{[double] Minima of the scale}

\item{beta}{[double] Maxima of the scale}
}
\value{
[list(t,X,Y)] Dataset Y, with covariate X and time axis t
}
\description{
Generate a test dataset, Y is a Gaussian law where location and scale move with time, X is the location and the scale parameter
}
\examples{
## Data
size = 2000
data = Dataset2(size)
t = data$t  ## Time axis
X = data$X  ## Covariate
Y = data$Y  ## Dataset to fit
}
