% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPDLaw.R
\docType{class}
\name{GPDLaw}
\alias{GPDLaw}
\title{GPDLaw (Generalized Pareto)}
\format{\code{\link{R6Class}} object.}
\usage{
GPDLaw
}
\arguments{
\item{loc}{[vector]
Location parameters}

\item{scale}{[vector]
Scale parameters}

\item{shape}{[vector]
Shape parameters}

\item{scale_cov}{[matrix]
Scale covariate for fit}

\item{shape_cov}{[matrix]
Shape covariate for fit}

\item{use_phi}{[bool]
Use exponential function as link function for scale}

\item{method}{[string]
Optimization method, default "BFGS"}

\item{verbose}{[bool]
Print warning and error message}
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class to generate, fit, and use a generalized pareto law.
}
\section{Methods}{

\describe{
  \item{\code{new(loc,scale,shape,use_phi,method,verbose)}}{Initialize pareto law with code{GPDLaw}}
  \item{\code{rvs(size,loc,scale,shape)}}{Random number generator}.
  \item{\code{density(x,loc,scale,shape)}}{Density}.
  \item{\code{cdf(Y,loc,scale,shape)}}{Cumulative distribution function.}.
  \item{\code{icdf(p,loc,scale,shape)}}{Inverse of Cumulative distribution function.}.
  \item{\code{sf(Y,loc,scale,shape)}}{Survival function.}.
  \item{\code{isf(p,loc,scale,shape)}}{Inverse of survival function.}.
  \item{\code{fit(Y,loc,scale_cov,shape_cov)}}{Fit the GPD law}.
}
}

\examples{
## Data
size = 2000
data = SDFC::Dataset2(size)
t = data$t
X = data$X
Y = data$Y

## Quantile regression for loc parameters
ltau = base::c( 0.05 , 0.95 )
qr = SDFC::QuantileRegression$new( ltau )
qr$fit( Y , X )
## Yq[,1] is the lower loc of GPD, and Yq[,2] the upper loc.
Yq = if( qr$is_success() ) qr$predict() else NULL

## Upper GPD
gpdU = SDFC::GPDLaw$new()
gpdU$fit( Y , loc = Yq[,2] , scale_cov = X )
Yu = gpdU$rvs(size)
print( gpdU$scale_coef_ ) ## Scale coef fitted
print( gpdU$shape_coef_ ) ## Shape coef fitted

## Lower GPD
gpdL = SDFC::GPDLaw$new()
gpdL$fit( -Y , loc = -Yq[,1] , scale_cov = -X ) ## Minima of Y is the maxima of -Y
Yl = -gpdL$rvs(size)
}
\keyword{datasets}
