% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{LMoments1}
\alias{LMoments1}
\title{LMoments1}
\usage{
LMoments1(Y)
}
\arguments{
\item{Y}{[vector] Dataset}
}
\value{
[lmom1] L-Moments of order 1
}
\description{
Compute the L-Moments of order 1 (just the mean...)
}
\examples{
## Data
size = 2000
data = Dataset0(size)
lmom1 = SDFC::LMoments1(Y)
}
