% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{LMoments2}
\alias{LMoments2}
\title{LMoments2}
\usage{
LMoments2(Y)
}
\arguments{
\item{Y}{[vector] Dataset}
}
\value{
[lmom2] L-Moments of order 2
}
\description{
Compute the L-Moments of order 2 (half mean of pairwise difference)
}
\examples{
## Data
size = 2000
data = Dataset0(size)
lmom2 = SDFC::LMoments2(Y)
}
