% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalLaw.R
\docType{class}
\name{NormalLaw}
\alias{NormalLaw}
\title{NormalLaw (Gaussian)}
\format{\code{\link{R6Class}} object.}
\usage{
NormalLaw
}
\arguments{
\item{loc}{[vector]
Location parameters}

\item{scale}{[vector]
Scale parameters}

\item{loc_cov}{[matrix]
Location covariate for fit}

\item{scale_cov}{[matrix]
Scale covariate for fit}

\item{use_phi}{[bool]
Use exponential function as link function for scale}

\item{method}{[string]
Optimization method, default "BFGS"}

\item{verbose}{[bool]
Print warning and error message}
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class to fit a Normal law with covariates
}
\section{Methods}{

\describe{
  \item{\code{new(use_phi,method,verbose)}}{Initialize normal law with code{NormalLaw}}
  \item{\code{fit(Y,loc_cov,scale_cov)}}{Fit the Normal law}.
}
}

\examples{
## Data
size = 2000
data = SDFC::Dataset2(size)
t = data$t
X = data$X
Y = data$Y

## Normal Law
norm = SDFC::NormalLaw$new()
norm$fit( Y , loc_cov = X , scale_cov = X )
print(norm$loc_coef_) ## Location coef fitted
print(norm$scale_coef_) ## Scale coef fitted

## In fact, it is better to fit:
norm_best = SDFC::NormalLaw$new()
norm_best$fit( Y , loc_cov = X[,1] , scale_cov = X[,2] )
}
\keyword{datasets}
