% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileRegression.R
\docType{class}
\name{QuantileRegression}
\alias{QuantileRegression}
\title{QuantileRegression}
\format{\code{\link{R6Class}} object.}
\usage{
QuantileRegression
}
\arguments{
\item{ltau}{[vector]
Vector of quantiles where we want to perform the fit}

\item{method}{[string]
Method used to fit, currently, only "Frish-Newton" is available}

\item{verbose}{[bool]
Print warning and error message}
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class to perform a Quantile Regression with covariates
}
\section{Methods}{

\describe{
  \item{\code{new(ltau,method,verbose)}}{Initialize Quantile Regression with code{QuantileRegression}}
  \item{\code{fit(Y,X)}}{Fit the quantile regression}.
  \item{\code{predict()}}{Return the quantile fitted}.
  \item{\code{coef()}}{Return coefficients fitted}.
  \item{\code{is_fitted()}}{TRUE if fit is already called}.
  \item{\code{is_success()}}{TRUE if fit is a success}.
  \item{\code{is_unfeasible()}}{TRUE if fit is not feasible}.
}
}

\examples{
## Data
size = 2000
data = SDFC::Dataset2(size)
t = data$t
X = data$X
Y = data$Y

## Quantile regression for loc parameters
ltau = base::seq( 0.05 , 0.95 , 0.01 )
qr = SDFC::QuantileRegression$new( ltau )
qr$fit( Y , X )
Yq = if( qr$is_success() ) qr$predict() else NULL 
}
\keyword{datasets}
